code clean up, structs for spike event and windows

+++ processing/lib/Biquad.h
@@ -39,7 +39,7 @@ public:
     void setFc(double Fc);
     void setPeakGain(double peakGainDB);
     void setBiquad(int type, double Fc, double Q, double peakGainDB);
-    float process(float in);
+    double process(double in);
     
 protected:
     void calcBiquad(void);
@@ -50,7 +50,7 @@ protected:
     double z1, z2;
 };
 
-inline float Biquad::process(float in) {
+inline double Biquad::process(double in) {
     double out = in * a0 + z1;
     z1 = in * a1 + z2 - b1 * out;
     z2 = in * a2 - b2 * out;
